Add grid and its automatic creation Add user input for grid size
Created a responsive grid that uses flexbox to keep its grid rows and squares inside, and responsively resize them as well.
Also added a button for the user to enter a custom grid size. Upon entering the size, the current grid resets and creates a new one with the correct size.